argv.py: fix numbers

+++ shell/argv.py
@@ -28,8 +28,8 @@ if __name__ == "__main__":
     # meaning that the length of system arguments will be
     # more than one, when command-line arguments exist.
     if len(sys.argv) > 2:
-            num1 = long(sys.argv[1])
-            num2 = long(sys.argv[2])
+            num1 = float(sys.argv[1])
+            num2 = float(sys.argv[2])
     else:
             print("This command takes two arguments and adds them")
             print("Less than two arguments given.")
